Modularize app with constants, helpers and Streamlit class
