+++ Project_work_Gruppo_4/app.py
@@ -1,0 +1,4 @@
+from src.rag_app import StreamlitApp
+
+if __name__ == "__main__":
+    StreamlitApp().run()
