+++ Project_work_Gruppo_4/src/rag_app/__init__.py
@@ -1,0 +1,9 @@
+from .pipeline import RAGPipeline
+from .extraction import TemplateExtractor
+from .streamlit_app import StreamlitApp
+
+__all__ = [
+    "RAGPipeline",
+    "TemplateExtractor",
+    "StreamlitApp",
+]
